requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
assemblyai==0.24.0
attrs==23.2.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.4
django-environ==0.11.2
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.0.3
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.45
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.17.0
orjson==3.10.0
packaging==23.2
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.2.3
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-transcript-api==1.0.3
